FE - refactoring/styles in your own folder
