+++ frontend/src/components/FormInput.jsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import "./formInput.css";
+import "../styles/formInput.css";
 
 const FormInput = (props) => {
   const [focused, setFocused] = useState(false);
