+++ frontend/src/index.css
@@ -4,14 +4,6 @@ code {
     monospace;
 }
 
-
-/* new */
-/* *{
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-} */
-
 .dFlex{
    display: flex;
    justify-content: center;
@@ -22,23 +14,6 @@ code {
 body{
   width: 100%;
   background-color: gray;
-}
-
-
-.wrapper{
-  margin: 150px auto;
-  width: 80%;
-}
-
-
-.product_price{
-  margin: 50px auto;
-  text-align: left;
-}
-
-.product_img{
-  margin: 0;
-  padding: 2em;
 }
 
 .mt-auto{
