+++ frontend/src/pages/Home.js
@@ -2,7 +2,7 @@ import React, {useEffect, useState, useContext } from 'react';
 import { useNavigate } from 'react-router-dom'
 import ProductService from '../services/ProductService';
 import { ShopContext } from '../context/shop-context';
-import "./home.css";
+import "../styles/home.css";
 
 const Home = () =>{
