+++ frontend/src/pages/Product.js
@@ -1,6 +1,8 @@
 import React, {useEffect, useState } from 'react';
 import ProductService from '../services/ProductService';
 import {useParams} from 'react-router-dom';
+import "../styles/product.css";
+
 
 const Product = () => {
     const [product, setProduct] = useState(null)
@@ -20,17 +22,17 @@ const Product = () => {
             return (
                     <div className='wrapper'>
                          <div className='product_img'>
-                            <img src={product.thumbnail}  style={{width: "500px", height: "500px"}}/>
+                            <img className='img_size'  src={product.thumbnail} />
                         </div>
                         <div>
                             <h2>{product.title}</h2>
                             <p>{product.description}</p>
-                            <h2>Price: <span className='price'> &nbsp;&nbsp;{product.price} ,- EUR</span></h2>
+                            <h2>Price: <span class="product-price">&nbsp;&nbsp;{product.price} ,- EUR</span> </h2>
                         </div>
                     </div>   
             )
         }
-         
+
     }
 
     return(
