+++ frontend/src/styles/home.css
