+++ frontend/src/styles/order.css
@@ -51,8 +51,6 @@
     width: 100%;
   }
 
-  .buy-button {
-    position: absolute; /* nastavíme absolútnu pozíciu pre tlačidlo */
-    padding: 10px 20px; /* pridáme padding pre tlačidlo */
-    margin: 20px;
+  .size-btn  {
+    width: 200px;
   }
