--- frontend/src/validations/EmailValidation.js
@@ -1,9 +1,0 @@
-
-const emailValidator = require('email-validator');
-
-function validateEmail(email) {
-  const isValid = emailValidator.validate(email);
-  return isValid;
-}
-
-module.exports = { validateEmail };
